# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Repaire Phone DZ - E-commerce API for phone repair equipment
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── BRANDS ─────────────────────────────────────────────────────────────────
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{productId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/coupon:
    post:
      operationId: applyCoupon
      tags: [cart]
      summary: Apply coupon to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponApplyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ─── WISHLIST ────────────────────────────────────────────────────────────────
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    patch:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── BANNERS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──────────────────────────────────────────────────────────────────
...
    Category:
      type: object
      required: [id, name, slug, sortOrder]
      properties:
        id:
          type: number
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        iconName:
          type: ["string", "null"]
        parentId:
          type: ["number", "null"]
        sortOrder:
          type: number
        productCount:
          type: number
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        iconName:
          type: string
        parentId:
          type: ["number", "null"]
        sortOrder:
          type: number
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        iconName:
          type: string
        sortOrder:
          type: number

    # ── BRAND ─────────────────────────────────────────────────────────────────
...
    CartItem:
      type: object
      required: [productId, name, price, quantity, images]
      properties:
        productId:
          type: number
        name:
          type: string
        price:
          type: number
        comparePrice:
          type: ["number", "null"]
        quantity:
          type: number
        images:
          type: array
          items:
            type: string
        stock:
          type: number
    Cart:
      type: object
      required: [items, subtotal, discount, shipping, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        discount:
          type: number
        couponCode:
          type: ["string", "null"]
        couponDiscount:
          type: number
        shipping:
          type: number
        total:
          type: number
        itemCount:
          type: number
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: number
        quantity:
          type: number
    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: number
    CouponApplyInput:
      type: object
      required: [code]
      properties:
        code:
          type: string

    # ── ORDER ─────────────────────────────────────────────────────────────────
...
    Coupon:
      type: object
      required: [id, code, discountType, discountValue, isActive, usageCount]
      properties:
        id:
          type: number
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: ["number", "null"]
        maxUses:
          type: ["number", "null"]
        usageCount:
          type: number
        isActive:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: number
        expiresAt:
          type: string
    CouponUpdate:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: number
        isActive:
          type: boolean
        expiresAt:
          type: string

    # ── BANNER ────────────────────────────────────────────────────────────────
...