# lib/api-zod/src/generated/types/couponApplyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Repaire Phone DZ - E-commerce API for phone repair equipment
 * OpenAPI spec version: 0.1.0
 */

export interface CouponApplyInput {
  code: string;
}